Restructure transfer handler to accept spinto transfer events
