+++ src/core/constants/PintoBase.ts
@@ -10,7 +10,8 @@ import {
   PINTO_WETH,
   WELL_STABLE2,
   POOL_TOKENS,
-  TOKEN_INFOS
+  TOKEN_INFOS,
+  SILOED_BEAN
 } from "./raw/PintoBaseConstants";
 import { beanDecimals, PoolTokens, Token, WellFnInfo } from "./RuntimeConstants";
 import { BI_10 } from "../utils/Decimals";
@@ -19,6 +20,10 @@ import { BI_10 } from "../utils/Decimals";
 
 export function getProtocolToken(): Address {
   return BEAN_ERC20;
+}
+
+export function getSiloBeanToken(): Address {
+  return SILOED_BEAN;
 }
 
 export function getProtocolFertilizer(): Address | null {
