+++ src/core/constants/raw/PintoBaseConstants.ts
@@ -3,6 +3,7 @@ import { PoolTokens, Token } from "../RuntimeConstants";
 
 // Protocol tokens
 export const BEAN_ERC20 = Address.fromString("0xb170000aeeFa790fa61D6e837d1035906839a3c8");
+export const SILOED_BEAN = Address.fromString("0x7F6A8E8bA3E4cFd92fA1e5Dcd89Ba8F7eF92aA12");
 export const PINTO_WETH = Address.fromString("0x3e11001CfbB6dE5737327c59E10afAB47B82B5d3");
 export const PINTO_CBETH = Address.fromString("0x3e111115A82dF6190e36ADf0d552880663A4dBF1");
 export const PINTO_CBBTC = Address.fromString("0x3e11226fe3d85142B734ABCe6e58918d5828d1b4");
