+++ src/subgraphs/beanstalk/manifests/ethereum.yaml
@@ -385,7 +385,7 @@ dataSources:
         - handler: handleExploit
           filter:
             kind: once
-      file: ../src/handlers/BeanHandler.ts
+      file: ../src/handlers/TokenHandler.ts
   ###
   # SILO
   ###
@@ -1036,7 +1036,7 @@ dataSources:
       eventHandlers:
         - event: Transfer(indexed address,indexed address,uint256)
           handler: handleTransfer
-      file: ../src/handlers/BeanHandler.ts
+      file: ../src/handlers/TokenHandler.ts
   - kind: ethereum/contract
     name: Bean-PreReplant-Exploit
     network: mainnet
@@ -1057,7 +1057,7 @@ dataSources:
       eventHandlers:
         - event: Transfer(indexed address,indexed address,uint256)
           handler: handleTransfer
-      file: ../src/handlers/BeanHandler.ts
+      file: ../src/handlers/TokenHandler.ts
   ###
   # FARM BALANCE
   ###
