+++ src/subgraphs/beanstalk/manifests/no-apy.yaml
@@ -52,7 +52,7 @@ dataSources:
         - handler: handleExploit
           filter:
             kind: once
-      file: ../src/handlers/BeanHandler.ts
+      file: ../src/handlers/TokenHandler.ts
   ###
   # SILO
   ###
@@ -703,7 +703,7 @@ dataSources:
       eventHandlers:
         - event: Transfer(indexed address,indexed address,uint256)
           handler: handleTransfer
-      file: ../src/handlers/BeanHandler.ts
+      file: ../src/handlers/TokenHandler.ts
   - kind: ethereum/contract
     name: Bean-PreReplant-Exploit
     network: mainnet
@@ -724,7 +724,7 @@ dataSources:
       eventHandlers:
         - event: Transfer(indexed address,indexed address,uint256)
           handler: handleTransfer
-      file: ../src/handlers/BeanHandler.ts
+      file: ../src/handlers/TokenHandler.ts
   ###
   # FARM BALANCE
   ###
